key attribute removed

+++ aspnetapp/Model/DtmtestTable.cs
@@ -11,7 +11,7 @@ namespace aspnetapp.Model
         public int? FkId { get; set; }
         public string TestString { get; set; }
         public DateTime? InsertDate { get; set; }
-        public decimal? UploadId { get; set; }
+        public int? UploadId { get; set; }
         public string UploadName { get; set; }
         public string UploadFile { get; set; }
         public string UploadFileName { get; set; }
